Torque Multiplier Fixes for Engine and Exhaust
Sorry, I forgot to include this earlier. These changes are required to ensure that torque and power output with the stock ecu, engine block, exhaust, and turbos are as close to reality as possible. It shouldn't affect your custom configs, but will affect the base cars. However, this is easily fixed by adjusting the turbo waste-gate limit and start to match my config.